CryptoPkg/MbedTlsLib: Fix incompatible declaration of my_snprintf()
LTO builds do stricter type checking than non-LTO builds and linker
warns about this.  `size_t` isn't necessarily the same type as
`long long`.

+++ CryptoPkg/Library/MbedTlsLib/Include/mbedtls/mbedtls_config.h
@@ -3573,6 +3573,8 @@
 // #define MBEDTLS_PLATFORM_NV_SEED_READ_MACRO   mbedtls_platform_std_nv_seed_read /**< Default nv_seed_read function to use, can be undefined */
 // #define MBEDTLS_PLATFORM_NV_SEED_WRITE_MACRO  mbedtls_platform_std_nv_seed_write /**< Default nv_seed_write function to use, can be undefined */
 
+#include <stddef.h>
+
 extern int
 my_printf (
   const char  *fmt,
@@ -3584,7 +3586,7 @@ my_printf (
 extern int
 my_snprintf (
   char        *str,
-  long long   size,
+  size_t      size,
   const char  *format,
   ...
   );
@@ -3592,7 +3594,6 @@ my_snprintf (
 #define MBEDTLS_PLATFORM_SNPRINTF_MACRO  my_snprintf
 
 #define MBEDTLS_PLATFORM_MEMORY
-#include <stddef.h>
 extern void *
 mbedtls_calloc (
   size_t  n,
